Move the class out of the common project into an appropriate place, and add additional constants for use in integration tests.
